src/libpcp/src/pmns.c: allow pmNameLookup() to be broken into multiple PDU round-trips
The size of the PMNS, qa/799 and recent pmlogger changes have conspired
to mean than pmlogger may gather a list of names for pmID lookup
that is so long that the resulting PDU sent to pmcd is > 64 Kbytes,
which triggers pmcd's DoS defence (ignoring the PDU and closing the
client connection).

In libpcp we now break the list of names into sublists and process
these in multiple round trips if required.