qa/check, qa/daily-cleanup: try to repair pmlogger_daily carnage
On VMs that are started just for QA runs we have:

1. it is unlikely that pmlogger_daily will have been run at the
   scheduled time, so there is a window between reboot and QA during
   which pmlogger_daily could be flogging itself to death, which has
   all sorts of bad interactions with QA ... e.g. qa/051 is a repeat
   offender for failing automated QA but passing when you come back
   to recheck it ... the failure is usually systemctl unable to start
   the primary pmlogger.

2. if there is a corrupted archive, or an achive with mismatched
   metadata, then pmlogextract will fail, which makes problem 1.
   even worse, and means it will never go away.  We make the chance of
   a metadata mismatch much higher when the MMV PMDA is used because
   in qa tests 213, 241 and 647 (that I've found so far) we associate
   PMID 70.1.1 with 3 different metric names (mmv.test.counter,
   mmv.mymmv.export_value and mmv.test.noinit) and different metric
   type and semantics.

3. no one is checking /var/log/pcp/pmlogger/pmlogger_daily.log for
   badness (this is unavoidable in this operational environment),
   meaning 2. is likely to be undetected

All of this was discovered with nuclear medicine during the autopsy
of the qa/051 cadaver.

The fix involves:

a. check always runs the (new) daily-cleanup script

b. daily-cleanup runs pmlogger_daily -N -V and looks to see if that
   script wants to run pmlogger_merge ... this means the scheduled
   pmlogger_daily run was not done or it errored

c. if pmlogger_daily needs to be (re)run, see if
   /var/lib/pcp/config/pmlogrewrite/pcpqa.conf exists (this file is
   not in the packages!) ... if it is not there, install it ... if
   it is there check if it matches the version in daily-cleanup and
   if not, update it

d. run pmlogger_daily if needs be, then check for all done using the
   pmlogger_daily -N -V recipe from step a.  if this fails wail loudly

Steps b. and c. and d. can generate "Info:" lines on stderr if
something needs to be done ... so don't be alarmed by these ... they
come from check, not the scripts check runs so there should be no QA
fallout in .out files.