logutil: add pmlogctl -m flag, to enable service migration
Only try and migrate new or matching processes to the farm
service if the -m/--migrate flag is given. Update the service
unit for pmlogger_farm_check.service to pass the -m flag to
pmlogctl.

This retains previous pmlogctl behavior - only enables the
new migration feature if the flag is given.

Resolves a bunch of QA pmlogctl group failures and qa/1204
is remade for the new flag.