pmlogger_check, pmlogger_daily, common.check: revert pmlc probe
We recently changed to something like
    if echo status | pmlc "$pid" 2>&1 | grep "^Connected ..."
but this has the disadvantage of requiring 3 select()'s at pmlogger
before the pmlc output is flushed.

Changing to
    if pmlc "$pid" </dev/null 2>&1 | grep "^Connected ..."
requires only 2 select()'s at pmlogger.

It makes a big difference if pmlogger is busy processing PDU round
trips to pmcd between the select()'s, especially if the network
latency is large.

This was make the "slow pmlogger startup" issues worse.

The simpler code is quicker and serves the same purpose to establish
that the pmlogger process has parsed the configuration file
successfully and is up and running.