pmdanvidia: updates to latest NVML, add a batch of new metrics
Several changes to get the nvidia metrics up-to-date with the
NVML library in the latest versions of CUDA.  The new metrics
include energy use, devices UUIDs (also as labels), supporting
both compute and graphics per-process stats APIs, as well as a
new per-process indom to integrate more easily with top-alike
tools like pcp-atop and pcp-htop.  A new option has been added
to the PMDA allowing an optional refresh interval to be given,
such that proactive sampling occurs - this optional technique
can be used to give client tools the opportunity to sample the
values over a much longer interval, calculating average values
over that time from the more rapid sampling in the PMDA.  This
technique is used within pcp-atop(1).

selinux updates to match current X server policy are included.

Resolves Red Hat BZ #1690590