pmlogger: Implement zeroconf defaults using an additional env file
Earlier, a new way provide zeroconf values for pmlogger was implemented in
d19714e27677e3148bb0d1ed113ca719a7459044 and
985920112a56ac4506ccc82b4033f745aedebeaf. This solution worked on systems with
systemd but not others. This caused problem for the piuparts test which
apparently does not start under systemd during the test[1].

To implement the solution in a more consistent way across all systems (with and
without systemd), the earlier solution is reverted with a new solution that
drops an additional defaults file with pcp-zeroconf package.

Tests:

- Install on a machine with non-systemd Debian based distribution. When pmlogger
is started, the PMLOGGER_INTERNAL is part in the pmlogger process environment
with value 10.

- Install on a Debian machine with systemd. When pmlogger is started, the
PMLOGGER_INTERNAL is part in the pmlogger process environment with value 10.

- Update /etc/default/pmlogger with a value for PMLOGGER_INTERNAL. That value is
used.

- Install without pcp-zeroconf. When pmlogger is started, the PMLOGGER_INTERNAL
is not part of pmlogger process environment.

- The file /etc/defaults/pmlogger_zeroconf is part of the pcp-zeroconf package.

Links:

1) https://piuparts.debian.org/sid/fail/pcp-zeroconf_5.3.4-1.log

Signed-off-by: Sunil Mohan Adapa <sunil@medhas.org>