Revert "qa/950: run the test without pmproxy -A to enable service advertising"
This reverts commit 7fe721c1dfd8fef633de305f864bde26d5575023.

My earlier change (which had not been pushed, so Mark had not seen it)
does effectively the same thing, but in common.avahi so it is available
for all tests that use avahi_service() (even though 950 is the only
one today).

The common.avahi change also brands $PCP_PMPROXYOPTIONS_PATH with
a comment to indicate it has been modified by a QA test, which is
just "good practice" when QA tests modify config files from package
installs.