pmlogger scripts: move zeroconf definitions to read-only location
Follow up on suggestion in #1462 and relocate the zeroconf primary
logger default sampling interval setting to a read-only location.
This file does not belong in any user-configurable location - the
clue is in the name '*zero*conf' - it's not *some*conf. ;-)  This
is a more FHS-compliant location for this file and makes it very,
very clear this is not a file that should ever be edited.