logutil/pmieutil: tweak pmlogger/pmie_farm output for invalid options
These two scripts just run pmlogger/pmie_check with all of
the arguments they're passed from the respective rc script.
They are not ment to be run manually by users.

In the QA environment with qa/1370, they're run with --help,
which is then blindly passed onto to the check script, which
fails with a usage message and non-zero exit status. The usage
message is thus for the check script, not the calling farm script,
which confused qa/1370 into thinking these commands support
all of those undocumented options.

So throw away the stderr err output, but exit 1 if the check
script fails. Outside of the qa environment, the err output
should be in the check script log anyway.