packaging: remove RHEL conditionals around {pmlogger,pmie_farm} service enablement
Remove the <= RHEL8 version conditionals around the enablement of the
pmlogger_farm and pmie_farm services in the %post scriptlet. With this
change, if the pmlogger service ends up enabled (either via upgrade or
by pcp-zeroconf), so does pmlogger_farm. Ditto for pmie / pmie_farm.

In a later release, e.g. PCPv6, we'll revisit this - the primary and
farm services should become independent.