libpcp_web: stop Redis initialization tasks if Redis version is unsupported
Previously the other tasks were run, but no response was received
(because Redis state was set to SLOTS_ERR_FATAL), and therefore the last
phase of the redisSlotsBaton which frees the baton was never called.

Also rename SERVER_VERSION to REDIS_VERSION to clarify which server
version this variable describes.