pmdabpf: support x86, arm and powerpc
This commit vendors vmlinux.h from kernel 5.14.0-22 on CentOS 9 Stream
for x86, arm and powerpc (for build environments without BTF or older
kernel versions), and adds a new configure option --with-pmdabpf-btf to
specify a BTF location. If this option is specified, a vmlinux.h will
be generated from this location, and this file will be used instead of
the vendored vmlinux.h file.