libpcp: add __pmDumpPDUTrace() and circular trace buffer in pdu.c
We keep track of the last "few" PDUs in and out, and __pmDumpPDUTrace()
may be used to dump out the history.  Scheme is similar to the "trace"
buffer that's been in pmcd since day one.

First target is the (relatively new) incidence of the message:
    Unexpected ERROR PDU from PMCD
that pmlogger is throwing now and then in QA.