pmproxy: fix race on teardown of pcp protocol proxy clients
There existed a race between pcp client socket teardown and
the corresponding proxied pmcd socket teardown.  Now we use
the existing client reference counting to ensure the events
are coordinated and client freeing is only done once.

Somce data structure cleanup is also done here.  It turns
out we do not use the pmproxy client doubly-linked list so
we can remove it completely - it was just a distraction in
this investigation.  This then allows simplification of one
mutex as its now only used for pending write handling (TLS)
and the per-client data structure is made slightly smaller.

This resolves an intermittent pmproxy failure in tests 1600
and 294 (and possibly others).

Fixes https://github.com/performancecopilot/pcp/issues/1480