logutil: add path handler for pmlogger_farm_check service
The pmlogger_farm_check service is responsible for checking pmlogger
farm processes, restarting any pmloggers that have failed and for
starting any new pmlogger farm processes after control file changes.

The service is driven by pmlogger_farm_check.timer, which starts
the service every 5 minutes. When the service is triggered by the
timer, it simply runs 'pmlogctl -m check' and then exits. See
pmlogctl(1) for details. This works fine but it does mean there may
be up to a five minute delay before any pmlogger control file changes
will be automatically applied to the pmlogger farm.

This commit adds a systemd path handler to the pmlogger_farm_check
service to also trigger the service when there are any changes to
$PCP_SYSCONF_DIR/pmlogger/control or to the control file directory
$PCP_SYSCONF_DIR/pmlogger/control.d (i.e. new control files are added
or changes have been made to any existing control files). The trigger
is applied only when an fd for the file or directory is closed (not
when it is opened or written to).

The existing timer is retained - to periodically check/migrate/restart
any dead loggers, regardless of whether there have been any control
file changes.

This has been manually tested but will need a new QA test - TODO.