pmdanvidia: fix mishandling of zero-byte size passed to realloc
Picked up during QA of recent nvidia changes - some hardware lacks
support for per-process metrics, or the hardware (GPU) has not yet
been accessed by a process using its resources, which had the side
effect that a zero-byte size argument was passed into realloc.  In
turn, this passes back something that can be freed and an issue in
the logic meant this would happen on subsequent calls also.

Resolves the QA failure and Red Hat BZ #2029301