pmlogger, pmie scripts: fix Before= lines in systemd files
The Before= lines in pmie_farm and pmlogger_farm files lacked
a link to the associated timer.  The BindsTo= lines had links
to the unrelated primary pmlogger/pmie_check timers.

Finally, remove a hard-coded path from these .in files, which
was inconsistent with the rest of the systemd files setup.