Resolve inconsistencies in new 'farm' systemd timers
When the farm systemd timers were introduced the check interval
was drastically reduced from half hourly to 5 minutely.  There
wasn't any discussion about rationales for this and its now not
consistent (does not dovetail at all) with the primary pmlogger
and pmie service.  If startup takes a long time (large farms or
slow networks) these will likely overlap constantly, and timing
should be such that we work with the primary services in mind.

Reset to half hourly for these checks, and lets revisit this in
the new year when the other systemd changes are being proposed.

Related to https://github.com/performancecopilot/pcp/pull/1495