src/pmlogger/src/callback.c: small changes for derived metrics
When a derived metric is logged, we need to fiddle with the PMID
so that it does not look like a derived metric when the archive is
replayed.

The "modified" PMID needs to appear consistently in all the
pmResults and the metadata ... the latter was not quite right
for the optional help text associated with a derived metric,
and right by accident for any labels associated with a derived
metric.

This commit addresses both issues.

Also qa/352 is expanded to exercise the logging of help text and
labels for derived metrics.