qa: bpf: don't run tests if kernel BTF is not enabled
Previously we also checked for a BTF section in
/boot/vmlinux-$(uname -r)
however with Ubuntu 20.04 (as used in GitHub Action) this proved to be
not sufficient for running our bpf programs (kernel too old).

The availability of the kernel BTF is a better indicator whether
the kernel support for eBPF is mature enough for our eBPF programs.