pmlogger_daily and pmlogger_check: more systemd-driven changes
systemd on vm07 (Debian 11.1) insists on running pmlogger_daily.service
as soon as pmlogger.service is started, even though we're expecting
pmlogger_daily.service to only be run from pmlogger_daily.timer at
00:10 (by default).

This breaks the behaviour that has been in place with the cron-driven
scheme from day one.  And in particular causes an archive rotation via
pmlogger "re-exec" as soon as pmlogger has started ... this doubles
the pmlogger startup cost, but also breaks a number of QA tests.

The problem may be present on other platforms, I've just done the
deep triage on Debian 11.

The "fix" is really quite messy (more kin to digging another defensive
ditch and filling it with sharpened stakes than a fix), and invloves:

- change the pmlogger_daily logic so that the re-exec is NOT done
  if the current archive name contains today's date (this means
  the 00:10 execution of the script is the only time the re-exec is
  likely to be done and returns us to the historical state of affairs)

- add -Z flag to pmlogger_daily to FORCE the re-exec if possible
  (needed for at least one QA test)

- add -z flag to pmlogger_daily to SKIP re-exec (needed for at least
  one QA test)

- be more chatty with -V in pmlogger_daily when SIGUSR2 is, or is not,
  sent to trigger re-exec

- reworking the SaveLogs scheme so that the log file names are (a) in
  ascending timestamp order, and (b) have nanosecond timestamp
  precision in their names (pmlogger_check and pmlogger_daily)

- augment the SaveLogs scheme so that the script arguments and
  the process geneology (via pstree if available) as appended to the
  log ... this helps diagnose who called a particular script and with
  which parameters (pmlogger_check and pmlogger_daily)