delta indom for V3 archives - part 2
Tweaks to the write-path, but more importantly this commit implements
the lazy "un-delta-ing" of the indoms on the read-path.

pmdumplog -i now works correctly for V3 archives with delta indom
encoding of the metadata, and since the code to support this is in
libpcp, all other archive reading clients should be OK now.