src/pmdumplog/pmdumplog.c: band-aid for Coverity CID 374710
I think this is really a false positive, as the only (normal) way
that an on-disk indom instance would _not_ have a namelist[] entry
is for TYPE_INDOM_DELTA records (which the code already handles).

But a corrupted archive just might get us here, so add a check
and "Botch" if the condition arises.