pmlogextract: refactor to use __pmTimestamp
This commit changes all the internal pmTimeval uses (except for
<mark> records, see below) and most of the struct timeval data
to use __pmTimestamp and all internal timestamp manipulation and
calculation in pmlogextract uses the higher resolutuon and Y2038-safe
data structure.

The pmTimeval struct still appears on the I/O paths for V2 archives,
but is filled from a __pmTimestamp before writing and is copied to
a __pmTimestamp after reading.

Still some work to be done when __pmResult changes flow thru libpcp
and pmlogger, at which point we'll have real V3 archives ... this
involves all the pmResult uses for the "data" volumes, plus all the
<mark> record support (both of which remain in V2 archive land).