pmlogextract: add -V <version> support
Command line option allows default version rule (for output
archive) to be gazumped.

Specificially for a V2 archive (myarchive), this:
    pmlogextract -V3 myarchive myarchive_v3
    pmlogmv myarchive_v3 myarchive
will (in the future) convert from V2 to V3 format.

Also qa/1377 (new) to exercise this and some explicit V2 and
V3 variants of the "foo" archive in qa/tmparch.