libpcp_archive: new library for apps that write PCP archives
libpcp_archive will be used to accumulate "helper" routines that are
useful for apps that need to be able to write PCP archives.

Historically there has been a lot of code duplication across pmlogger,
pmlogextract, pmlogreduce, pmlogrewrite, libpcp_import, ...  but the
introduction of "delta indom" for V3 archives means the amount of
potential duplication was about to explode.

So, taking PCP 6.0 as a good point in time to tidy this mess up.

Initially (in this commit) all of the build and packaging changes
are in place for libpcp_archive, plus pulling some useful routines
out of pmlogger and moving them to libpcp_archive.