assorted: delta indom rethink
I started building man pages for the libpcp_archive routines and
realized there was a fatal flaw in the "delta indom" plan of record,
namely it was not possible to delete (or drop) an instance with
internal instance id of zero.

Quick rethink and change of plan ....
    On disk, for a deleted instance in a delta indom record, instead
    of instlist[i] being < 0, we leave instlist[i] to be the real
    internal instance identifier and encode a strindx[i] value of -1,
    which becomes namelist[i] == NULL when the indom is loaded, then
    change all the "is it a deleted instance" guards to use namelist[i]
    == NULL instead of instlist[i] < 0.

This produced changes in:
- man pages
- libpcp
- libpcp_archive
- pmlogger
- pmdumplog
- qa

There may be others if there is additional fallout from full QA runs.

Fortunately noone but me and QA is using V3 archives yet.