libpcp, libpcp_import, utils: transition to using __pmResult
This transitions a number of additional places relating to fetch
in libpcp over to using the internal __pmResult structure, which
is timestamp-size agnostic.  Places include fetch, fetcharchive,
interp.c, logfetch.c, log import, derived metrics, and so on.

The fetch path now automatically transitions to using the highest
resolution timestamps available internally, by detecting a remote
pmcd highres feature bit, and sending the highres PDU types if it
is supported.

Many tools with deep knowledge of PCP internals are affected by
this change; in particular pmlogger and many of the pmlog* tools,
pmlc, dbpmda and even a few obscure parts of pmie (pmdasummary).