Big refactor for InDom data moving to/from archive metadata files
Historically, we used pmInResult to encode indom data between pmdas and
pmcd, and for PDUs between pmcd and clients.

The same data structure was then also used for indom data moving to and
from archives.  But the V3 archives work, and in particular the "delta"
indom external format changes, have exposed some deficiencies in pmInResult
for code that needs to read and/or write archive metadata records.

Note that most clients are insulated from all of this because they read
existing archives via the PMAPI routines that hide all of the details
of archive encoding.  So the code that is exposed to this is pmlogger
(to write the archives) and the applications that read (and possibly
write) archives directly, e.g. pmlogextract, pmlogrewrite, pmlogreduce,
pmdumplog, and some QA apps.

This commit introduces __pmLogInDom_io as a replacement for pmInResult
on the archive I/O code paths.  In additon to the numinst, instlist[]
and namelist[] fields (from pmInResult), __pmLogInDom_io includes the
pmInDom and the __pmTimestamp.

Note: unfortunately we all ready have a __pmLogInDom data structure
      that is used for InDom data that has been loaded from an archive
      and linked into a mess of hashed data structures that are
      used within libpcp, hence the _io ugliness in the suffix for
      __pmLogInDom_io (better naming suggestions would be welcome)

The changes fall into 2 main camps:

1. using __pmLogInDom_io in place of pmInResult
2. argument refactoring to use __pmLogInDom_io instead of some subset
   of numinst, instlist[], namelist[], indom and stamp as arguments

Either change may also require an additional #include for <pcp/libpcp.h>
to get the __pmLogInDom_io typedef.

API changes

[] pmaDeltaInDom(), pmaSameInDom() and pmaSortInDom(): all in
   libpcp-archive and <pcp/archive.h>, arguments of type *pmInResult
   changed to *__pmLogInDom_io
[] __pmLogEncodeInDom(): was using an earlier version of __pmLogInDom_io
   (the prototype for these changes), so a small change to use the current
   version
[] __pmLogPutInDom(): arguments pmInDom indom, __pmTimestamp *tsp,
   int numinst, int *instlist, char **namelist all dropped and replaced
   by __pmLogInDom_io *lidp
[] __pmLogAddInDom(): arguments __pmTimestamp *tsp and pmInResult *in
   replaced by __pmLogInDom_io *lidp
[] addindom() (defined in internal.h so not visible outside libpcp)
   arguments pmInDom indom, __pmTimestamp *tsp, int numinst, int *instlist
   and char **namelist all dropped and replaced by __pmLogInDom_io *lidp
[] __pmLogLoadInDom(): arguments pmInResult *inp and __pmTimestamp *tsp
   replaced by __pmLogInDom_io *lidp

API caller changes
[] qa/src/chkputlogresult2.c: for __pmLogPutInDom()
[] src/libpcp_import/src/archive.c: for __pmLogPutInDom()
[] src/libpcp_web/src/discover.c: __pmLogAddInDom()
[] src/pmdumplog/pmdumplog.c: for __pmLogLoadInDom()
[] src/pmlogextract/pmlogextract.c: for __pmLogLoadInDom()
[] src/pmlogger/src/callback.c: change __localLogGetInDom() to use
   __pmLogInDom_io and changes for __pmLogPutInDom() and __pmLogAddInDom()
[] src/pmlogger/src/logue.c: for __pmLogPutInDom()
[] src/pmlogreduce/indom.c: for __pmLogPutInDom()
[] src/pmlogrewrite/indom.c: for __pmLogLoadInDom() and __pmLogPutInDom()

Other changes
[] qa/src/scanmeta.c: just seems more natural to use __pmLogInDom_io
   rather than pmInResult when unpacking physical indom records from the
   metadata file