pmlogger: resolve uninitialized __pmLogInDom_io field use (covscan)
Coverity picked up a pmaDeltaInDom call which used an uninitialized
field (timestamp) in the struct pointed to by the second parameter.
I suspect this is benign because we properly initialize this field
before writing it out, but as uninit'd variable access is undefined
behaviour to the compiler let's ensure the struct is fully setup.