libpcp + apps: change API for __pmEncodeResult
Add a __pmLogCtl pointer arg; NULL for PM_CONTEXT_HOST else this
will allow us to know the archive version number.

This is a preparatory step to the larger work of unifying the pmResult
encode and decode code paths, which are complicated because they have
a PDU and V2 archive variant and a V3 archive variant.