pmdadenki: fix crash in labels callback with multiple instances
Remove some leftovers from the initial creation of this PMDA based
on pmdasimple.  The labels callbacks in particular could encounter
a sigsegv when accessing beyond a static global data structure.  A
couple of commented-out calls used the 'now' indom from simple but
that no longer exists.

Also updated the labels to reflect metrics with units in watt hours
and watts, as per Christian's note.

Resolves https://github.com/performancecopilot/pcp/issues/1504