pmdamssql: fix config file logic, python string handling
Python strings do not translate by casting to appropriate
boolean equivalents - this has to be done carefully, and
manually.  This problem causes a bad interaction between
the ansible-pcp mssql role and pmdamssql when setting the
trusted connection flag, several possible values of this
entry in the config file (incl. no value at all) would as
a result equate to True when False would have been a much
more appropriate setting.  Testing with various flavours
of local mssql connection setup via qa/1337 shows that we
are in a much healthier place now with this commit.

Several debug statements on frequently-called paths also
were observed to be causing unneccessary string expansion,
only for the result to be thrown away (non-debug modes) -
a waste of resources.  These have been commented out.

Resolves Red Hat BZ #2057615