V3 archives: final on-disk format changes
Specifically, 32-bit nsec for timestamps in data (__pmResult)
records and no padding.

Lots of code changes here with internal API changes in for libpcp
(in libpcp.h, not pmapi.h), removal of a lot of *HighRes* alternate
code (the pmResult Encode and Decode paths now handle all of the PDU,
V2 archives and V3 archives variants of the format).

These changes get all of QA passing on bozo (except for 1829, there
is still a problem in pmlogrewrite V2 -> V3 conversion to be tracked
down).