pmlogger: drop old PDU buffer decoding for diagnostics
With the rearranging of __pmResults and PDU buffer handling to
support pmlogger writing either version 2 or 3 archives, we no
longer need to arrange special decoding of pinned PDU buffers.
This means the leakage of on-disk pmResult structure internals
can be dropped from pmlogger.  The same information can be got
from the pmPrintResult in the pmDecode[HighRes]Result calls.

Resolves https://github.com/performancecopilot/pcp/issues/1545