src/pmlogger/src/fetch.c: cull PDU buffer scraping
The only reason this file includes "knowledge" about the PDU structure
is to support -Dfetch diagnostics ... turns out this is a really
bad idea (a) because it always was a bad idea, (b) because it is now
totally broken on 32-bit platforms (as in qa/038 dies horribly).

Chuck all this code away ... there is a better way to provide the
diagnostics and I'll return to that in a later commit.