libpcp: high resolution log label and time window APIs
Provide new APIs to access the v3 log labels start timestamp
in high resolution form, as well as optional zoneinfo field,
in the form of pmGetHighResArchiveLabel(3).  The context end
of archive timestamp is made available in high resolution too
via a new pmGetHighResArchiveEnd routine.

Update the pmGetOptions routine to support a v3 ABI such that
PMAPI client tools can opt-in to working with nanosecond time
throughout.  pmval is converted as a proof of concept.

Resolves https://github.com/performancecopilot/pcp/issues/1372
Resolves https://github.com/performancecopilot/pcp/issues/1374
Resolves https://github.com/performancecopilot/pcp/issues/1377