qa/check: run tests under timeout(1) control if possible
Suggestion from Andreas Gerstmayr.

To allow us some control, opt for SIGABRT to signal the timeout and
install a signal handler in common.rc (that everyone sources).

The timeout is set to 5m(ins) for each test script execution, which
should be plenty (almost twice as long as the longest running test
on bozo today).

If the timeout happens, we'll see in .out.bad something along the
lines of:
    Aborted (core dumped)
    [may be repeated several times]
    !!!
    !!! Killed from check by timeout ...
    !!!