src/libpcp_web/src/discover.c: fix __pmLogLoadInDom use in pmDiscoverDecodeMetaInDom()
For 32-bit platforms (like vm21 where a bunch of pmproxy tests were
failing with a pmproxy SEGV), the return value from __pmLogLoadInDom()
needs to be checked more carefully ... if the result is 1, then
lid.namelist[] is really a re-use of part of buf[].

If the __pmLogInDom_io struct is going to be passed on to someone
who might later free() parts of it, then it is necessary to allocate
a new namelist and initialize all the elements, namelist[i] =
lid.namelist[i], and then make lid.namelist point to the new
namelist.

The problem only happens on 32-bit platforms where sizeof(int) ==
sizeof(void *), and the buf[] slots are reused.  For 64-bit platforms
lid.namelist[] is already malloc'd.

Another Akmal-driven premature optimization comes home to bite 25+
years on!