Makepkgs: fix for Debian build (and Qt packaging failure in CI)
We run configure twice for a Debian build.  This change involves the
first invocation where we were not correctly picking the environment
variables from debian/rules that need to be set to make the first
configure behave like the second one (that is run inside the Debian
package building ecosystem).  The commit fixes that.

After the first configure and before the second configure we run
debian/GNUmakefile create debian/control that is used in the real
build.

The specific problem was that
	configure_tools = export DEBUG=-DNDEBUG QT_SELECT=5;
was not in play for the first configure, critically QT_SELECT=5 was
not set.  We now roll the dice ... and depending on what Qt version(s)
you have installed, we may or may not set ENABLE_QT that drives the
whole packaging of pmchart, pmtime, pmdumptext, et al.

In the QA Farm I have Qt4 and Qt5 on my Debian 10 system (vm11)
... I win.

In CI, the Debian 10 system has only Qt5 installed ... we
all lose.