src/selinux/make-type-conditional: (new) helper script
As I said in the previous commit, everytime we bump into a new selinux
type, class, attribute, whatever that is not defined on some platform
we're committed to building on, we have to go through a convoluted
circus of changes to configure.ac, builddefs.in, pcpupstream.te.in,
GNUmakefile, GNUlocaldefs and qa/917 to fix it up and make type,
(or class or attribute, ...) conditional

I am _really_ tired of this whole bucketload of maintenance that seems
(to me) to be of marginal value (at best).  So this scripts automates
the process subject to these constraints:

- only works for types
- only works when there is one rule involving the types
- does not try to fix qa/917, as that one's a different can of worms
  and not well suited to awk/sed munging

Since types are often the cause of issues here (way more common than
classes and the the other selinux metadata types are extremely rare),
and a priori the types causing the issues are unusual (the common
ones are supported everywhere, or we've already handled them), the
"only one rule" restriction is OK.

Even if this script does not get it 100% correct, it will be a lot
closer and less error-prone than the manual editing we've been doing
previously.