pmie systemd: add missing systemd configuration and shell code
During the addition of the pmlogger/pmie farm systemd setup at
the end of last year, parts that were done for pmlogger didn't
end up being done for pmie.  This resulted in incorrect cgroup
setup until the check script (timer-based) arrives to fix it.

This commit also fixes a benign bug in the migrate_pid_cgroup
function where a return code was mishandled and early-exit was
not performed on a loop-over-possible-paths.  Only noticed it
because it incorrectly reported an error in verbose mode - its
unlikely anyone ever could notice this in practice, but its a
little bit more efficient now.

Resolves https://github.com/performancecopilot/pcp/issues/1522