libpcp_archive: add new pmaPutMark() function
Knowledge about V2 and V3 formats is leaking out in assorted parts
of the code and is not always correct.

Most recently I've been trying to fix up <mark> record handling
pmlogextract which was broken in some important corner cases.

pmaPutMark() hides all the version and format details and provides a
higher-level interface for those PMAPI clients that need to be able
to append a <mark> record to an archive's current data volume.