man/man3i/new: helper script
I'm trying to encourage more of the important non-PMAPI routines to
benefit from man pages ... this helper script starts with a function
name, finds the function declaration in the source (so we have a full
prototype and know which library it is in), builds a skeletal but
customized man page, does the debian packaging work (often forgotten in
the past, leading to build breakage), and adds the new man page to git.

The recently added pmaPutMark(3) man page was created with ./new in
the man3i subdir.

If this proves useful it could be the model for similar scripts
in other sections of the man pages, and/or extended for additional
libraries.