libpcp & libpcp_archive: retire pmaPutMark, beef up __pmLogWriteMark
The API for pmaPutMark() is tidier, but we already had __pmLogWriteMark()
(that I had not noticed when I created pmaPutMark()).

So this commit drops all trace of pmaPutMark() from libpcp_archive
and refines the API for __pmLogWriteMark() in libpcp, and makes the
necessary changes in all places that called either routine in the
past.

For extra points, there is now a __pmLogWriteMark(3) man page.