qa: refactor test 1466 for increased reliability
Instead of using 'pmsleep 5' (racey on very slow systems) as
the primary protagonist in this test, use pmpause which does
not exit until we force it to.  Drop most of the sleeps from
the test with this, increase diagnostics, and see if this is
enough to get CI back on track with this test.