libpcp: implement a pmHighResStore(3) interface in the PMAPI
Thought we would not need this but have found programs that
issue pmFetch (pmHighResFetch) then want to pmStore based on
the result.  In this case, its much more convenient is we've
got a highres variant of store (even though the timestamp is
ignored).

We'll use this in the python API and updated tests shortly.