qa/494: force timezone for babylon.pmview_v2 archive
On ci-centos9 PCP 6.0.0 x86_64 CentOS Stream9 in CI this test (and
a bunch of others are failing because the timezone is not being
set correctly).

Try forcing TZ=PST8PDT9 (compared to PST8PDT from the archive label).
These two are equivalent, so if this works it will identify the bug
in CentOS Stream9.