rc-pmcd: reorder the systemd paths attempted in migrate_pid_cgroup
An earlier bug in this code resulted in all paths being attempted
before bailing out with an acceptable path - try reordering these
paths so that the unified path in particular is moved back into a
higher priority position in the list.  Hopefully this will tackle
the remaining CI failure when this is run inside containers there.