libpcp & libpcp_pmda: rename some HighRes functions
On reflection, it would appear we could be a little more consistent
with the naming of the "HighRes" variants of some library functions.

Most are unaltered, but we're making these changes:
    pmHighResFetch => pmFetchHighRes
    pmHighResFetchArchive => pmFetchHighResArchive
    __pmHighResFetchLocal => __pmFetchHighResLocal
    pmHighResStore => pmStoreHighRes
    pmHighResSortInstances => pmSortHighResInstances
    pmHighResSetMode => pmSetModeHighRes
    pmdaEventHighResAddParam => pmdaEventAddHighResParam
    pmdaEventHighResGetAddr => pmdaEventGetHighResAddr

Since many of these are new in PCP 6.0, the change is a simple
substitute (in all the places the old name appeared), but for a
few of them the function is in a public interface that has already
escaped in a release, and so the old name is preserved ... in the
header files, in the libraries, in the Python wrappers and the man
pages, but the implementation is a thin wrapper around a call to the
newly named function.

Ones in this latter group are pmHighResFetch, pmdaEventHighResAddParam
and pmdaEventHighResGetAddr.