qa/src/torture_indom.c: small fixes for indom operations
Because an archive indom is timestamped, pmGetInDom() is really
only going to work _after_ at least one successful pmFetch() for the
associated metric.

So,
- add pmSetMode() for archives so we use PM_MODE_FORW which sort
  of forces the pmFetch() to succeed, even if the requested metrics
  are not in the first archive log record
- move the pmFetch() to come before the pmGetInDom() call

This had been a non-issue previously, because the only qa tests using
torture_indom are either using (a) live hosts, or (b) archives with
simple and regular logging, so the indom and the metric are both
defined at the start of the archive.  I was playing with some other
archives, like qa/archives/omnibus, and this exposed the logic flow
in the test program.