qa/252: rework ... very fragile size calculations here
This test is critically sensitive to the size of the pmResult in an
archive and omitted to account for the size of the epilogue pmResult
at the end of the archive.

Also dropped sample.lights from the pmlogger config because the strang
value of this metric may change, and that changes the pmResult length.

Not broken for V3 archives per se, but V3 archives pushed the size
calculations from stable to unstable.